<refactor> barrel files pipe
