+++ frontend/src/app/pages/cliente/solicitacao/visualizar-solicitacao/visualizar-solicitacao.component.ts
@@ -6,7 +6,7 @@ import { SolicitacaoService,EquipamentoService,OrcamentoService } from "../../..
 import { CommonModule } from "@angular/common";
 import { ModalComponent } from "../../../../components/modal/modal.component";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
-import { EstadoAmigavelPipe } from "../../../../shared/pipes/estado-amigavel.pipe";
+import { EstadoAmigavelPipe } from "../../../../shared/pipes";
 import { HistoricoUtils } from "../../../../shared/utils";
 
 @Component({
