+++ frontend/src/app/pages/empregado/solicitacao/visualizar-solicitacao/visualizar-solicitacao.component.ts
@@ -6,7 +6,7 @@ import { SolicitacaoService,EquipamentoService,FuncionarioService,OrcamentoServi
 import { CommonModule } from "@angular/common";
 import { ModalComponent } from "../../../../components/modal/modal.component";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
-import { EstadoAmigavelPipe } from "../../../../shared/pipes/estado-amigavel.pipe";
+import { EstadoAmigavelPipe } from "../../../../shared/pipes";
 import { HistoricoUtils } from "../../../../shared/utils";
 import { NgxCurrencyDirective } from "ngx-currency";
